[hotfix] Rename all *end-to-end/itcases* to *e2e*
