+++ statefun-e2e-tests/statefun-sanity-e2e/src/main/protobuf/verification-messages.proto
@@ -1,0 +1,73 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+syntax = "proto3";
+
+package org.apache.flink.statefun.e2e.sanity;
+option java_package = "org.apache.flink.statefun.e2e.sanity.generated";
+option java_multiple_files = false;
+
+/*
+ * A command is addressed to a specific target funcction and triggers some action by that target.
+ * Commands can be nested to an arbitrary depth.
+ */
+message Command {
+    FnAddress target = 1;
+
+    oneof kind {
+        Noop noop = 2;
+        Send send = 3;
+        Modify modify = 4;
+    }
+}
+
+/*
+ * Sent by functions to egress after modifying its state.
+ */
+message StateSnapshot {
+    FnAddress from = 1;
+    int32 state = 2;
+}
+
+/*
+ * Target function address of commands.
+ */
+message FnAddress {
+    int32 type = 1;
+    string id = 2;
+}
+
+/*
+ * Modify the function's state by adding @delta to it's state.
+ */
+message Modify {
+    int32 delta = 1;
+}
+
+/*
+ * Send 1 or more commands to different recipients.
+ */
+message Send {
+    repeated Command commandToSend = 2;
+}
+
+/*
+ * A dummy command, does not require any action by the recipient.
+ */
+message Noop {
+}
