+++ statefun-e2e-tests/statefun-sanity-e2e/src/test/resources/flink-conf.yaml
