+++ statefun-e2e-tests/statefun-sanity-e2e/src/test/resources/log4j.properties
