--- statefun-end-to-end-tests/statefun-sanity-itcase/src/main/java/org/apache/flink/statefun/itcases/sanity/FnCommandResolver.java
@@ -1,96 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.statefun.itcases.sanity;
-
-import static org.apache.flink.statefun.itcases.sanity.Utils.toSdkAddress;
-
-import org.apache.flink.statefun.itcases.sanity.generated.VerificationMessages.Command;
-import org.apache.flink.statefun.itcases.sanity.generated.VerificationMessages.FnAddress;
-import org.apache.flink.statefun.itcases.sanity.generated.VerificationMessages.Modify;
-import org.apache.flink.statefun.itcases.sanity.generated.VerificationMessages.Noop;
-import org.apache.flink.statefun.itcases.sanity.generated.VerificationMessages.Send;
-import org.apache.flink.statefun.itcases.sanity.generated.VerificationMessages.StateSnapshot;
-import org.apache.flink.statefun.sdk.Address;
-import org.apache.flink.statefun.sdk.Context;
-import org.apache.flink.statefun.sdk.FunctionType;
-import org.apache.flink.statefun.sdk.annotations.Persisted;
-import org.apache.flink.statefun.sdk.match.MatchBinder;
-import org.apache.flink.statefun.sdk.match.StatefulMatchFunction;
-import org.apache.flink.statefun.sdk.state.PersistedValue;
-
-/**
- * Simple stateful function that performs actions according to the command received.
- *
- * <ul>
- *   <li>{@link Noop} command: does not do anything.
- *   <li>{@link Send} command: sends a specified command to another function.
- *   <li>{@link Modify} command: increments state value by a specified amount, and then reflects the
- *       new state value to an egress as a {@link StateSnapshot} message.
- * </ul>
- */
-public final class FnCommandResolver extends StatefulMatchFunction {
-
-  /** Represents the {@link FunctionType} that this function is bound to. */
-  private final int fnTypeIndex;
-
-  FnCommandResolver(int fnTypeIndex) {
-    this.fnTypeIndex = fnTypeIndex;
-  }
-
-  @Persisted
-  private final PersistedValue<Integer> state = PersistedValue.of("state", Integer.class);
-
-  @Override
-  public void configure(MatchBinder binder) {
-    binder
-        .predicate(Command.class, Command::hasNoop, this::noop)
-        .predicate(Command.class, Command::hasSend, this::send)
-        .predicate(Command.class, Command::hasModify, this::modify);
-  }
-
-  private void send(Context context, Command command) {
-    for (Command send : command.getSend().getCommandToSendList()) {
-      Address to = toSdkAddress(send.getTarget());
-      context.send(to, send);
-    }
-  }
-
-  private void modify(Context context, Command command) {
-    Modify modify = command.getModify();
-
-    final int nextState =
-        state.updateAndGet(old -> old == null ? modify.getDelta() : old + modify.getDelta());
-
-    // reflect state changes to egress
-    final FnAddress self = selfFnAddress(context);
-    final StateSnapshot result =
-        StateSnapshot.newBuilder().setFrom(self).setState(nextState).build();
-
-    context.send(Constants.STATE_SNAPSHOT_EGRESS_ID, result);
-  }
-
-  @SuppressWarnings("unused")
-  private void noop(Context context, Object ignored) {
-    // nothing to do
-  }
-
-  private FnAddress selfFnAddress(Context context) {
-    return FnAddress.newBuilder().setType(fnTypeIndex).setId(context.self().id()).build();
-  }
-}
